Refactor: melhorando e otimizando o codigo
