+++ src/App.tsx
@@ -1,14 +1,12 @@
-import { Box, ChakraProvider, Grid, theme } from "@chakra-ui/react";
-import { ColorModeSwitcher } from "./components/color-mode-switcher/ColorModeSwitcher";
+import { Box, ChakraProvider, theme } from "@chakra-ui/react";
+import { ColorModeSwitcher } from "./components/ColorModeSwitcher/ColorModeSwitcher";
 import { AppRoutes } from "./routes/AppRoutes";
 
 export const App = () => (
   <ChakraProvider theme={theme}>
-    <Box textAlign="center" fontSize="xl">
-      <Grid minH="100vh" p={3}>
-        <ColorModeSwitcher justifySelf="flex-end" />
-        <AppRoutes />
-      </Grid>
+    <Box position="fixed" left="0" top="0" right="0" textAlign="right">
+      <ColorModeSwitcher justifySelf="flex-end" />
     </Box>
+    <AppRoutes />
   </ChakraProvider>
 );
